do not creaty npy if exists

+++ nnunetv2/training/dataloading/utils.py
@@ -92,9 +92,11 @@ def _convert_to_npy(npz_file: str, unpack_segmentation: bool = True, overwrite_e
     try:
         a = np.load(npz_file)  # inexpensive, no compression is done here. This just reads metadata
         if overwrite_existing or not isfile(npz_file[:-3] + "npy"):
-            np.save(npz_file[:-3] + "npy", a['data'])
+            if not os.path.exists(npz_file[:-3] + "npy"):
+                np.save(npz_file[:-3] + "npy", a['data'])
         if unpack_segmentation and (overwrite_existing or not isfile(npz_file[:-4] + "_seg.npy")):
-            np.save(npz_file[:-4] + "_seg.npy", a['seg'])
+            if not os.path.exists(npz_file[:-4] + "_seg.npy"):
+                np.save(npz_file[:-4] + "_seg.npy", a['seg'])
     except KeyboardInterrupt:
         if isfile(npz_file[:-3] + "npy"):
             os.remove(npz_file[:-3] + "npy")
